Move follow count visibility into the Config system
Instead of a column in the `Identity` model, this configuration now
belongs to the `Config` model.

I also created a new Privacy page under the Identity settings to put
this new configuration into.
