+++ templates/identity/create.html
@@ -14,7 +14,6 @@
             {% include "forms/_field.html" with field=form.domain %}
             {% include "forms/_field.html" with field=form.name %}
             {% include "forms/_field.html" with field=form.discoverable %}
-            {% include "forms/_field.html" with field=form.visible_follows %}
         </fieldset>
         <div class="buttons">
             <button>Create</button>
