+++ templates/identity/view.html
@@ -72,7 +72,7 @@
             </table>
     {% endif %}
 
-    {% if identity.visible_follows %}
+    {% if identity.config_identity.visible_follows %}
       <div class="stats">
         <ul>
           <li><strong>{{ following_count }}</strong> following</li>
