+++ templates/settings/_menu.html
@@ -6,6 +6,9 @@
     <a href="{% url "settings_interface" %}" {% if section == "interface" %}class="selected"{% endif %} title="Interface">
         <i class="fa-solid fa-display"></i> Interface
     </a>
+    <a href="{% url "settings_privacy" %}" {% if section == "privacy" %}class="selected"{% endif %} title="Privacy">
+        <i class="fa-solid fa-shield"></i> Privacy
+    </a>
     <h3>Account</h3>
     <a href="{% url "settings_security" %}" {% if section == "security" %}class="selected"{% endif %} title="Login &amp; Security">
         <i class="fa-solid fa-key"></i> Login &amp; Security
