+++ templates/settings/profile.html
@@ -10,7 +10,6 @@
             {% include "forms/_field.html" with field=form.name %}
             {% include "forms/_field.html" with field=form.summary %}
             {% include "forms/_field.html" with field=form.discoverable %}
-            {% include "forms/_field.html" with field=form.visible_follows %}
         </fieldset>
         <fieldset>
             <legend>Images</legend>
