+++ tests/users/views/settings/test_privacy.py
@@ -1,6 +1,7 @@
 import pytest
 from pytest_django.asserts import assertContains, assertNotContains
 
+from core.models.config import Config
 from users.models import Follow
 
 
@@ -10,6 +11,7 @@ def test_stats(client, identity, other_identity):
     Tests that follow stats are visible
     """
     Follow.objects.create(source=other_identity, target=identity)
+    Config.set_identity(identity, "visible_follows", True)
     response = client.get(identity.urls.view)
     assertContains(response, "<strong>1</strong> followers", status_code=200)
 
@@ -19,7 +21,6 @@ def test_visible_follows_disabled(client, identity):
     """
     Tests that disabling visible follows hides it from profile
     """
-    identity.visible_follows = False
-    identity.save()
+    Config.set_identity(identity, "visible_follows", False)
     response = client.get(identity.urls.view)
     assertNotContains(response, '<div class="stats">', status_code=200)
